After a jump to another file, also re-query the jumps, because the jumplist got updated with the text for the jumps, whereas it previously only contained the buffer name
Thanks to Daniel Hahler for sending a patch.
Especially in small terminals, jump messages may not fit and cause a hit-enter prompt. Truncate messages in s:Echo().
Local jump message only considers the header, but not the file jump messages. If its one, and cmdheight is 1, add its width to the number of reserved columns, as we append the following location. Thanks to Daniel Hahler for the patch.
The warning message before a remote jump isn't truncated to fit.